requirements.txt
flask==2.3.3
pymongo==4.6.0
flask-cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0
passlib==1.7.4
flask-bcrypt==1.0.1
email-validator==2.1.0
pytest==7.4.0
flask-mail==0.9.1
matplotlib==3.8.0
pandas==2.1.1
openpyxl==3.1.2
reportlab==4.0.4
seaborn==0.12.2
numpy==1.24.3
